requirements.txt
azure_storage==0.37.0
matplotlib==3.8.0
meteostat==1.6.7
numpy==1.26.4
pandas==2.2.0
pyzmq==25.1.0
scikit_learn==1.3.2
SQLAlchemy==2.0.21
